requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
pydantic-settings
python-jose[cryptography]
supabase
httpx
email-validator
slowapi
